+++ packages/expo-calendar/android/src/test/java/expo/modules/calendar/EventRecurrenceTest.kt
@@ -2,7 +2,6 @@ package expo.modules.calendar
 
 import expo.modules.calendar.domain.event.records.RecurrenceRuleEntity
 import expo.modules.calendar.domain.event.records.input.RecurrenceRuleInput
-import expo.modules.kotlin.apifeatures.EitherType
 import expo.modules.kotlin.types.ConvertedValue
 import expo.modules.kotlin.types.Either
 import expo.modules.kotlin.types.IncompatibleValue
@@ -25,7 +24,6 @@ class EventRecurrenceUtilsTest {
     TimeZone.setDefault(null)
   }
 
-  @OptIn(EitherType::class)
   @Test
   fun testCreateRecurrenceRule_withAllFields() {
     val opts =
@@ -120,7 +118,6 @@ class EventRecurrenceUtilsTest {
   }
 }
 
-@OptIn(EitherType::class)
 private inline fun <reified L : Any, reified R : Any> eitherWithFirst(value: L): Either<L, R> {
   val types = listOf(typeOf<L>(), typeOf<R>())